preventing deleting categories and tags if they are attached to posts

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -68,6 +68,9 @@ class CategoryController extends Controller
     public function destroy(string $id)
     {
         $category = Category::find($id);
+        if ($category->posts->count()) {
+            return redirect()->route('categories.index')->with('error', 'Ошибка! У категории есть записи!');
+        }
         $category->delete();
         return redirect()->route('categories.index')->with('success', 'Категория удалена');
     }
