+++ app/Http/Controllers/Admin/PostController.php
@@ -2,8 +2,8 @@
 
 namespace App\Http\Controllers\Admin;
 
-use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Http\Controllers\Controller;
 use App\Models\Post;
 use App\Models\Tag;
 use Illuminate\Http\Request;
@@ -13,6 +13,8 @@ class PostController extends Controller
 {
     /**
      * Display a listing of the resource.
+     *
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
      */
     public function index()
     {
@@ -22,6 +24,8 @@ class PostController extends Controller
 
     /**
      * Show the form for creating a new resource.
+     *
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
      */
     public function create()
     {
@@ -32,6 +36,9 @@ class PostController extends Controller
 
     /**
      * Store a newly created resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function store(Request $request)
     {
@@ -53,23 +60,28 @@ class PostController extends Controller
         return redirect()->route('posts.index')->with('success', 'Статья добавлена');
     }
 
-
     /**
      * Show the form for editing the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
      */
-    public function edit(string $id)
+    public function edit($id)
     {
         $post = Post::find($id);
         $categories = Category::pluck('title', 'id')->all();
         $tags = Tag::pluck('title', 'id')->all();
-
         return view('admin.posts.edit', compact('categories', 'tags', 'post'));
     }
 
     /**
      * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  int  $id
+     * @return \Illuminate\Http\RedirectResponse
      */
-    public function update(Request $request, string $id)
+    public function update(Request $request, $id)
     {
         $request->validate([
             'title' => 'required',
@@ -82,7 +94,9 @@ class PostController extends Controller
         $post = Post::find($id);
         $data = $request->all();
 
-        $data['thumbnail'] = Post::uploadImage($request, $post->thumbnail);
+        if ($file = Post::uploadImage($request, $post->thumbnail)) {
+            $data['thumbnail'] = $file;
+        }
 
         $post->update($data);
         $post->tags()->sync($request->tags);
@@ -92,12 +106,17 @@ class PostController extends Controller
 
     /**
      * Remove the specified resource from storage.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\RedirectResponse
      */
-    public function destroy(string $id)
+    public function destroy($id)
     {
         $post = Post::find($id);
         $post->tags()->sync([]);
-        Storage::delete($post->thumbnail);
+        if ($post->thumbnail) {
+            Storage::delete($post->thumbnail);
+        }
         $post->delete();
         return redirect()->route('posts.index')->with('success', 'Статья удалена');
     }
